feat: support twu proposal team member persistence from front-end to back-end
